requirements.txt
asgiref==3.3.3
Django==3.2
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-environ==0.4.5
django-mathfilters==1.0.0
djangorestframework==3.12.4
environ==1.0
model-bakery==1.2.1
mysqlclient==2.0.3
pytz==2021.1
sqlparse==0.4.1
